pyproject.toml
[project]
name = "parseltongue-claude"
version = "0.1.0"
description = "MCP server wrapping Parseltongue's symbolic verification engine for Claude Code"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Nissim Lebovits" }]
keywords = ["mcp", "parseltongue", "verification", "claude", "llm"]

dependencies = [
    "mcp>=1.0.0",
    "parseltongue-dsl>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/parseltongue_claude"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[project.scripts]
parseltongue-mcp = "parseltongue_claude.server:main"
